services/data-processing/preprocessor/requirements.txt
# API and web server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
starlette==0.27.0

# Database
pymongo==4.6.0
motor==3.3.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Storage
boto3==1.28.64
minio==7.2.0

# Data processing
pandas==2.1.2
numpy==1.26.1
scipy==1.11.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.0
pyarrow==14.0.1
dask==2023.10.1
polars==0.19.12

# Text processing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
markdown==3.5
python-docx==1.0.1
PyPDF2==3.0.1
pdfminer.six==20221105

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
tenacity==8.2.3
httpx==0.25.1
aiohttp==3.8.6
joblib==1.3.2

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
